requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
pandas==2.1.3
numpy==1.25.2
python-dotenv==1.0.0
google-generativeai==0.3.2
jinja2==3.1.2
aiofiles==23.2.0
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
lxml==4.9.3
webdriver-manager==4.0.1
